WB_AddDelta: Avoid asserting out on unknown delta operation modes
Since 05b89934 (Enhance the available delta modes in the wavebuilder
panel, 2015-03-26) we have different operation modes for the delta
values.

And in that commit it was decided that we assert out on unknown
operation modes. This is unwise as it completely breaks creating
stimsets with older MIES versions which not yet know about future
operation modes.

Instead we know complain and don't create the stimset.

Original-CommitID: 02ac54bfab52fb9e0fa3638fc33bbcf351b95f90