Add a stimset acquisition cycle ID
The repeated acquisition cycle ID was added to be able to distinguish
sweeps from different data takes.

But it turns out this grouping is too coarse as we want to operate on all
sweeps of a specific stimset cycle instead.

Therefore we introduce the new "Stimset acquisition cycle ID". This is
an arbitrary value which is written into the labnotebook.

This ID changes if one of the folllowing properties change:
- RAC
- stimset
- setimset checksum
- set cycle count

To keep track of the changes we generate a fingerprint from them and
compare that to the one of the previous sweep.

An "easier" way to solve that quest and using the fingerprint as ID does
not work as we want to have different IDs for the case where the same
stimset is acquired on multiple headstages or the same stimset is
acquired in RA scenarios with Repeat Sets and Indexing.

Original-CommitID: 30586ec4668745397a9938ce7ad261afd865a7cb