Fix locked indexing bug
When the first stimset is not the one with the most sweeps we fail to
produce the correct acquisition order with locked indexing.

The reason is that we updated valdisp_DataAcq_SweepsActiveSet after
calling IDX_IndexingDoIt but we calculcated the number of sweeps before
calling it.

The proper fix is to update the value in the DAP_UpdateDAQControls
with REASON_STIMSET_CHANGE_DUR_DAQ as that is the correct place to do
it.

The changed code was introduced in 150e9122 (Repeated Acquisition: Major
cleanup, 2016-02-05) but the bug is present before as well.

Add a test to ensure that the behaviour stays correct.

Original-CommitID: 4741cdcb8bd503528bde62a8c1e02e31c0d76193