Analysis functions: Add PRE_SET_EVENT and fix POST_SET_EVENT for indexing
The analysis functions can not be used reliable with indexing as
POST_SET_EVENT is broken. In addition a PRE_SET_EVENT is missing, as for
indexing PRE_DAQ_EVENT is only called once.

Fix both issues. We currently determine the necessity for
PRE_SET_EVENT/POST_SET_EVENT in DC_PlaceDataInITCDataWave as that is the
only place we know the current column of the stimset and now how many
sweeps are in the stimset.

This will break with manually skipping sweeps. A future version with a
completely reworked indexing logic will fix that.

Original-CommitID: 053ee5c17275dd6f4c86321096090c4ac3cf4830