Fix active set count updating for unlocked indexing
We did not update the active set count for unlocked indexing. This is
functionality wise not yet a problem as we need it updated only for
locked indexing, but for a future application it is required to have it
correct for both modes.

Add also two small tests to verify the active set count.

Original-CommitID: 23b20d5e6921e80ff55079d7feffcbc9dd64f769