ED_createWaveNoteTags: Document the ITI if it changed
Up to now we are gathering all settings in DC_PlaceDataInITCDataWave and
then after the sweep was acquired write them to the labnotebook.

This approach fails for settings like ITI which are used after the sweep
was acquired but can also be changed during the sweep.

Now we document the value if it changed. We don't use the direct way of
writing into the sweepSettingsWave as that would overwrite the original
value.

We can't write the changed ITI in RA_RecalculateITI as that is called
after the POST_SWEEP_EVENT and that analysis function from that event
might add labnotebook entries which would hide our labnotebook block.
The reason is that we currently assume that all settings from a specific
entry source type (DATA_ACQUISITION_MODE, TEST_PULSE_MODE) are in a
continous range. And adding a DATA_ACQUISITION_MODE entry in
RA_RecalculateITI would break that assumption. See also the usage of
FindRange in GetLastSettingNoCache.

Original-CommitID: 437f1f609ea400cb253a7a5edf3887d9e3e517f4