PSQ_SquarePulse/PSQ_Rheobase: Jump to end of DAQ in POST_SET_EVENT
Previously we did not support indexing with these analysis functions.
But as we now support indexing we must jump to the end of DAQ on
failure. This was missed in these two cases, as we don't decide if the
set passed prior to executing POST_SET_EVENT.

Original-CommitID: 9c4613eda7feeaf724176adea5d2656bb12cba51