Wavebuilder: Add explicit delta values
Our current delta mode operations always require a functional dependecy
of the delta values. So arbirtrary delta values are not possible.

Introduce the explicit delta operation which allows the user to supply a
list of delta values. Note that the first entry of that list is added to
the second sweep and so on.

Original-CommitID: 4fd00838455bbdf43b61dbdaf1e36c9f184906a1