DC_PlaceDataInITCDataWave: Don't error out on zero cutOff for TP MD
Since a6a5203b (Bugfix: Change Testpulse to ITCDataWave conversion
logic, 2015-07-29) we set the rest of the ITCDataWave to zero for TP MD.

This makes sense as in nearly all cases the ITCDataWave will not be a
integer multiple of a single test pulse.

But for cases where this assumption does not hold, or with new hardware
which does not required that the ITCDataWave is larger than the real
data, this leads to an runtime error.

Original-CommitID: 0ec61d9d884dfe513253ee4a83d404eb38eefaa3