TP_ResetTPStorage: Don't redimension the original wave for storing it
When "Save TP record" is enabled we used to redimension the TPStorage
wave and then save it under a running number. This involves two memory
allocations and also results in arbitrary row sizes of TPStorage.

Use the RMD flag of Duplicate to avoid both issues. So now we also have
row sizes of a power of two.

Original-CommitID: 2854e2881ba27f71049194fc5f37deb54f847aea