DA_Ephys: Support per sweep ITI
We used to calculate the ITI as maximum of the ITI of all active sets or
sets in range with indexing turned on.

As we now can have per sweep ITIs in the stimsets the logic for
determining the ITI of a sweep changed. It is now the maximum ITI of all
sweeps of the active sets only.

Syncing from follower to lead panel is still done for yoking.

As the sweep is only known after configuring a device the ITI in the
panel is only updated once all devices are configured.

Original-CommitID: 6d481cad7c7c54488e027a40a275f3a91db18b48