AI_FillAndSendAmpliferSettings: Complain about missing but expected amplifier connection
A user report indicates that occasionally the amplifier settings of a
particular headstage could not be queried.

This is most likely due to the AI_SelectMultiClamp call failing. Either
because the MCC XOP/Application is buggy or the user closed the MCC App
midsweep.

To help debug that issue we add some messages.

Original-CommitID: 0da0512836739830147254164044755ce89f8f28