Documentation: Don't highlight verbatim blocks
According to [1] sphinx tries to be clever and highlights verbatim
blocks in the default highlightning language.

We only use verbatim for outputting text so we can just switch the
default highlightning language to text.

This also works for our code blocks as we use \rst blocks for that and
not the doxygen commands @code/@endcode.

[1]: https://github.com/michaeljones/breathe/issues/387

Original-CommitID: e48169d08b7cdc269284702647d858c971f40cd9