Call WBP_UpdateITCPanelPopUps from all stimset loading functions
Since 0f90e902 (DAP_PopMenuChkProc_StimSetList: Only search MenuExp for
normal wave controls, 2018-07-18) we don't call it anymore for all
controls (as that was a bug anyway). But now we end up in the situation
that the a newly configured MIES triggered the check added in
2aa54cab (DAP_CheckSettings: Complain if the calculated number of sweeps
in the set is zero, 2018-08-01).

So let's add updating the menuExp data, as this what
WBP_UpdateITCPanelPopUps does, to the stimset loading functions.

Original-CommitID: ab36ea856195db9a18585ad706e520c562dada52