Acquisition support for NI DAC devices in multi device mode
Data setup in DataConfiguratorITC supports NI devices. Internally
NI devices use 1D waves for each channel, while using the slightly
adapted ITCChanConfigWave as template.

NI devices use a separate TTL channel for each line (no bit combine)

NI devices use the AD clock as master clock source for DA and TTL.
DA and TTL hardware is setup and armed at PrepareAcq and stays on hold until
StartAcq starts the AD acquisition/AD clock.

Added dual paths for core hardware acquisition for ITC and NI devices.

Background tasks for TP and DAQ know NI devices now

DQM ActiveDeviceList has a field for the hardwareType of the device added

ActiveDevicesTPMD has a field for the hardwareType of the device added

Check for non existing FIFO when accessing FIFO for NI devices

This is merely a workaround for an issue that appears rarely,
where a FIFO is accessed that should still be present.
Because it is a workaround a message was added that is shown if
it appears.

Check for free disk space when starting the FIFO for NI devices

When a data acquisition is started there is a check for free disk space
now. The threshold is derived from the sample rate and the fifo size with
a safety factor.

FFT source wave for PowerSpectrum may not have any NaNs, adapted initialization
of OscilloscopeDataWave to fill DAC waves with zero

details:
The cause was that FFT does not allow the source wave to have any NaNs
even if the FFT is done column wise and a separated column contains NaNs.

With NI device support it was changed that the FFT source is the
OscilloscopeDataWave (ODW) which was initialized with NaN.
Now the DAC columns of the ODW are initialized to 0
while the ADC columns are still initialized to NaN (they are filled on
acquisition)

Added padding to the next power of two to FFT for Powerspectrum

details:
The run time of the FFT directly relates to the prime factor decomposition
of the size of the source wave. By padding the source wave to the size of the
next power of two the decomposition contains prime factors of 2 only.

Original-CommitID: 21a34705ed3d192ebfa289f6c31afa3b95816a97