HW_NI_IsRunning: Update for DAQ operations with NI hardware
DAQ operations run with ADC using the fifo method of NiDAQmx XOP to acquire
data. At the same time on the analog-in clock the DAC data is setup through
WaveForm output as well as the TTL waves output through DIO.
DAC and TTL is after setup always active since it is setup in infinite
repetition mode and armed.

The HW_NI_StopAcq functions stops all (ADC/DAC/TTL) activities. So it is
sufficient to check only for a single operation that is always active on
TP and DAQ. So the DAC activities are checked here. In case it was already
finished or an error occured the fDAQmx_WF_IsFinished function returns an
error code. Thus the error message has to be further checked to distinguish
between an actual error or the state that currently no activity is running.

Original-CommitID: 4124f9c328087c39f17d53fb5a4959af51138fe2