DA_Ephys: Ensure that the ITI is always reached for manual sweep starts as well
For repeated acquisition we wait after the sweep the required ITI. But
for manual sweep starts we did not do that.

Now we ensure that the ITI is always reached even for manual sweep
starts. This feature is user-configurable and turned off by default.

Original-CommitID: abfc3470994914771a8144d9c01336aa80a6f386