OOD_CreateStimSet: Speed it up
We are using FindLevels from the start of the acc wave to find the last
falling edge. But for that we can also use FindLevel starting from the
end, which is much faster.

Although we reverse the search direction, the edge types are still the
same as the documentation states

/EDGE=2: Searches only for crossing where the Y values are decreasing as
         level is crossed from wave start towards wave end.

Original-CommitID: 311fed9dfa361f2795a302b90452b72e381cf024