oodDAQ: Cache everything possible and nicify interface for callers
Since the initial implementation in d102c07d (Add new data acquisition
mode: Optimized overlap distributed acquisition, 2016-08-10) the oodDAQ
codes was not touched seriously.

The original motivation for only caching the results was that this is
the most expensive operation and that caching the regions and the
stimsets would be too much hassle or in the case of the stimsets requite
too much memory.

Now two years later it turns out that the delay experienced by the user
due to the additional region calculation and stimset creation is much
worse compared to the additional memory requirements.

Therefore we now cache the offset, regions and the offsetted stimsets.
To achieve that in a nice manner we introduce the new function
OOD_GetResultWaves which does all the internal offset calculation etc.
which formerly had to be done by the caller.

Original-CommitID: 647aab75bcd4c8b199cae13d95670059e6097351