CHI_CheckXOP: Add hash checks for NIDAQ XOP
In the future we don't want to include the NIDAQ XOP in the git
repository. In order to have a way to check that the user has the
correct XOP we now calculate the hash of the XOP and compare that to a
known value. Only if these hashes match do we return without error.

Original-CommitID: 1a5ae53e79f6412b03b808da9f920fdace9ed0df