Threaded FIFO handling: Increase timeout to 50ms
The TIMEOUT_IN_MS constant is used as timeout for ThreadGroupGetDFR in
TFH_FifoLoop and therefore gives the maximum elapsed time between runs.

The higher this timeout the fewer number of fifo values are in the
queue. This helps in the reading part of the queue as we only use the
latest value anyway but we currently don't have a way of flushing the
queue.

This change speeds up TP MD and makes the GUI more responsive but should
be deemded experimental.

Original-CommitID: 430589b06b4268c44e068eff3b537eb4c6497391