NeurodataWithoutBorders: Flush file after every sweep in separate thread
One of the worst cases for MIES and its users is losing valuable data.
So we would like to be always 100% sure that the acquired data is
written to disc. Unfortunately this can be quite expensive to do.

But since IP8 the HDF5 XOP is threadsafe so we can at least flush the
file from a preemptive thread after each sweep saving.

Original-CommitID: 41e8f2febee4b1eec862efa118415c6b1a442d17