HW_NI_CloseDevice: Kill the FIFO always
In case the device is running we kill the FIFO inside HW_NI_StopAcq, but
we also want to kill it if the device is not running.

This helps in debug & test cycles to get rid of a lingering FIFO. Due to
that change the FIFO might not be running so we need to test for that in
HW_NI_KillFifo as well.

Original-CommitID: 98d08c540ec617a810885d1b1ad7a733fb788f18