DC_MakeNITTLWave: Document TTL settings and rework it completely
Since NIDAQ hardware support we are not documenting the TTL channel
settings into the labnotebook.

Adding these settings required to rework DC_MakeNITTLWave completely as
we need to call it only once and in that step need to create all
labnotebook entries. The reason is that we can call
DC_CalculateChannelColumnNo only once for each active channel as it
carries a lot of internal state which is modified on each call.

Therefore we need to fill in TTLWave and create the LBN entries in one
step.

The new LBN entries are called "TTL channels" and "TTL stim sets". We
don't have a "TTL bits" entry as we have one bit per channel for NI
hardware.

Original-CommitID: 7e66b7b32d5046c4d43fb9155712ab8a991ec7ad