IH_KillTemporaries: Remove NWB history reference number
Since eca8f7bc (NWB: Store the Igor Pro history in NWB on interactive
export, 2018-05-29) we capture the history on first DA_Ephys panel
creation and store it then in the NWB file.

But if you load an existing PXP file and store the data into NWB an RTE
comes up as histRefNumber is invalid.

Delete histRefNumber on experiment close therefore. We don't delete it
in IH_KillTemporaries as that is called on save as well.

Original-CommitID: 6cf6c48cdb00888e049e4fcae590b6364008d2f5