DAP_GetHardwareType: Loosen algorithm and rename it
DAP_GetHardwareType uses currently the list of active NI devices
suitable for DAQ for determining if a device is of ITC/NI/UNKONWN type.

This is too fragile for e.g. saved experiments where the list of
connected NI devices is empty or if the devices are not connected any
more.

Resurrect the loose checking implemented in 32f2de8 (DAEphys panel lists
now valid NI DAC devices in the Devs popup menu, 2018-08-06) but extend
it as the deviceNumber must be empty for NI DAQ devices.

Original-CommitID: 941634fea606fce08de2cc9c3c53b32a46a7485d