GetAllDevices: Add support for NI hardware
We used to be able to know all device names for DAQ hardware. But for NI
hardware the name can be arbitrary.

We now get a list of folders and look into these if we can find
ITCDeviceIDGlobal which is a very strong hint that the folder belongs to
a once-locked device.

Original-CommitID: d77e48f54c4d9fddf19abb22dd73d97228d60662