Add Async Framework
This framework allows to push workloads asynchroneously
that are executed in a separate thread.

The procedure file includes a description as well as a code example.

- automatic in order execution of read outs
- data folder strucutre setup on prepared
- stop waits for read outs to finish within timeout
- DF push logic to/from threads centralized
- Thread DF is marked
- Thread DF setup can be done when frame work is stopped
- Globals from Free DFs defined in constants
- Stopping the frame work kills the Async DF

Tests for Async Frame Work

Several test for the Unit Testing Framework

Start and Stop
Setup
Run Workers without a specific order
Run Workers in a specific order
Workers that raise a runtime error

Added a limit of 64 threads as maximum in ASYNC_Start

This prevents creating a high number of threads which may
reduce system stability and response.

AsyncFW Test: Added test to start frame work with infinite number of threads

It should not be possible to start a high number of threads as they will
block the Igor Pro.

AsyncFW: runtime errors in readout functions are caught and assertion is thrown

AsyncFW Test: Added test TASYNC_RunErrorReadOut with a
failing readout function

AsyncFW: Added optional parameter to ASYNC_Start to disable Readout task

An optional parameter 'disableTask' was added to ASYNC_Start that allows
to disable the ReadOut task. ReadOut has to be performed by calling the
ASYNC_ThreadReadOut function.

AsyncFW Test: Add Test if Background task is not running when disabled

Async_Start allows to disable the creation of the ReadOut Background task.
The test TASYNC_StartWOTask checks if the task was not started.

Removed an unecessary check in TASYNC_Start_Stop

AsyncFW: Revised documentation for Async Framework

AsyncFW: Moved documentation to separate restructured text file

in /doc/asyncframework.rst

The separate file was included in index.rst as well as in MIES_Async.ipf

fix: for failing Test TASYNC_StopTimeOut in IP7

On ThreadGroupRelease IP7 seems to generate a virtual runtime error
This error can be caught however by AbortOnRTE / try catch endtry.
The fix interpretes an RT with code 0 as this virtual RT to end the thread.

fix: changed test ASYNC_TimeOutForce to have an infinite worker

The recent patch for IP7 introduced catching the rt triggered by
ThreadGroupRelease. The worker for ASYNC_TimeOutForce now catches
this RT, so it can continue until the end of time, at least until
it is killed forcefully by igor.

fix: access to NVAR <null> ref is checked now properly

with rtFunctionErrors=1 a runtime error is thrown when an access to
a NVAR refeence that is <null> happens. Previously the result was a
NaN with a silent conversion to it.

Since Igor 8 build 32615

Original-CommitID: 601010fcb05670b1689bb2b317ec0382284d0b93