HW_ITC_MoreData/HW_ITC_MoreData_TS: Fix return value for offset usage
Since 5b454452 (ITCChanConfigWave: Adapt to new ITC XOP requirements,
2018-03-08) MIES can utilize the new offset feature of the ITCXOP2.

The offset feature allows to start DAQ at a row larger than zero of
the data wave. It was a good design decision back then that the now
queried fifo position is relative to that offset, i.e. fifo position
zero is at the offset and *not* at row zero of the data wave.

Unfortunately we forgot to use that offset in determining when DAQ
should stop as stopCollectionPoint does not know anything about the
offset.

This bug can lead to OOB crashes when using large offsets which are
close to the stopCollectionPoint and with only one channel.

This bug can be reproduced with the following setup:
- 1 Headstage
- 10s sweep
- Set the offset manually to 1800000 in MIES_DataConfiguratorITC.ipf.
  (Search the line "ITCChanConfigWave[][%Offset] = 0")
- Start DAQ

Now you see that only a small portion at the end is acquired. And then
you see a long time no change as DAQ still seems to be ongoing. That is
the bug!

Original-CommitID: 2751141d822afcc736ed5d3ec569f54ab61eb911