fix: vc_redist package was not installed by the installer
The execute call notation required a separate notation for parameters.
The original call was non-working as soon as one added a parameter inside the
double quotes.

Also added a Sleep for the time between process ending and file close,
otherwise vc_redist.x64.exe was not deleted (and wouldn't be deleted on
uninstall later as well)

Original-CommitID: edcbbafafe88d5cda34dd823410a733285189871