AB_ResetSelectionWave: Handle zero rows for listbox waves
If you scan a folder with no pxps/nwbs in it we end up with no entries
in the listbox, but this case is not handled properly in
AB_ResetSelectionWave().

Bug present since the initial introduction in b419fd07 (Add the first
version of the Analysis Browser package, 2015-02-25).

Original-CommitID: d17543bf51f0620bddcc8c0aa9db00a0d652a226