ANALYSIS_BROWSER_SUPP_VERSION: Raise to latest EXPERIMENT_VERSION
In 211499e4 (Raise EXPERIMENT_VERSION, 2018-08-09) we upgraded the
experiment version to 2 because NI hardware support was added.

The folder ITCDevices was renamed to HardwareDevices for NI hardware
addition.

We now translate, depending on the pxpVersion, the folder names in
AB_TranslatePath.

We can not, as we do in MIES, rename the datafolder itself as we need to
reload data via LoadData and we just can not rename a folder in a PXP.

Original-CommitID: c65e0603cac63b93399f18ca717f5ed273fcfc30