Cache: Create the statistics wave with the correct size
Experiments prior to 370fcaf4 (On Quit and New: Flush the cache,
2018-03-01) still have the cache in the PXP. And experiments older than
ab795b55 (Cache: Add statistics for each entry, 2018-03-23) don't have
the stats wave.

What now happens if you load experiments that old is, that the cache wave
is created due to the wave getters magic kicking in, but the values/key
waves might have been already resized.

We therefore have to ensure that we create it with the correct size to
begin with.

Of course we can not fill the stats wave with data as that is not
available.

Original-CommitID: 9994de92a0dfa214d85eddc8d2d4eed673369d2e