fix CalculateTPLikePropsFromSweep totalOnsetDelay issue
totalOnsetDelay was included in the base (100%) for calculating the 90 % range
for the base line average as well as for the elevated average.
So the range start for base line averaging was
90 % of (totalOnsetDelay + base line fraction) and for elevated range
90 % of (totalOnsetDelay + base line fraction + active tp duration).

The range is now changed to be
for base line avg: 10 % of base line fraction ending at first edge
for elevated avg: 10 % of active tp duration ending at second edge
Both endings are reduced by one point.

The edge locations are truncated to integer after determination such that
they point to the last point of the bas eline fraction as well as
to the last point of the active test pulse.

The full calculation of the ranges is now floating point.

Original-CommitID: 4278eec7ab5083cc93c54bd1108b3c1be2bacee1