IsValidObjectName: Restrict the name length to 31 bytes in IP8 as well.
As long as we are using IP7/8 in parallel we want to be able to load the
experiments in both versions.

Therefore we want to limit the valid name length to the IP7 limit.

This also affects analysis function parameter names.

Original-CommitID: 97cfefa05b3826b4b65270e50715a90fc1c899e4