upload-github-release-asset-helper.sh: Add last updated timestamp to release text
It is currently difficult for users to find out when the given release
was last updated.

Unfortunately github does not have a default way of setting the release
date, therefore we get the release properties and add a timestamp to the
description.

See https://developer.github.com/v3/repos/releases/#edit-a-release for
the properties.

Original-CommitID: eee1914a4e935c0ed7e208d356347f541b2a5cc1