upload-github-release-asset-helper.sh: Add last updated timestamp to release text
It is currently difficult for users to find out when the given release
was last updated.

Unfortunately github does not have a default way of setting the release
date, therefore we get the release properties and add a timestamp to the
description.

See https://developer.github.com/v3/repos/releases/#edit-a-release for
the properties.

Original-CommitID: 617423f42dee4da64809c7c7e30001739cf2fb67