WB_ParameterWvsNewerThanStim: Fix edge case when modification timestamps are equal
The wave modification timestamp has a granularity of 1 second.

If the following steps
- Paramter wave loading from NWB
- Stimset creation
- Parameter wave modification

takes less than 1 second the stimset recreation check will not be
successfull as the modification dates compare to equal.

In that case we use the newly added modification count values from the
stimset wave note to judge if we need to recreate it or not. We can not
only reyly on the wave modification counts as these start at zero on
wave creation.

Original-CommitID: c46e390e11e70c0a115d5f6863581490c2e88806