fix: Sweep SkipAhead resets to -1 in GUI on DAQ, following DAQ fails
changed param in DAP_SetVarProc_skipAhead to max(0, IDX_MinNoOfSweeps(sva.win) - 1)

The problem is to be related that the SetVar_DataAcq_skipAhead controls Limits are set to
High -1
Increment 0
Low 0

when calling DAP_ResetSkipAhead(panelTitle) in DC_CalculateChannelColumnNo
the generic function
PGC_SetAndActivateControl(panelTitle, "SetVar_DataAcq_skipAhead", val=0)
is called.

The value 0 is set by the generic function
PGC_SetAndActivateControl with respectLimits = 1.
Due to the behavior of limit in line 1114 in MIES_GUIUtilities high (-1)
is returned and put into the control.

For the next DAQ a sweep number of -1 is invalid and
it aborts on the setup of DAQ.

In MIES_DAEphys_Macro the control is initialized in line 3552 with
SetVariable SetVar_DataAcq_skipAhead,limits={0,0,1},value= _NUM:0
which is different to the limits state above.

The limits are changed to 0, -1, 1 when the associated control Proc
is executed with DAP_setSkipAheadLimit(sva.win, IDX_MinNoOfSweeps(sva.win) - 1)
 in line 4787 of MIES_DAEphys.ipf.

The parameter is not aware of the state where there is no sweep yet
acquired and IDX_MinNoOfSweeps(sva.win) returns 0.

Original-CommitID: 7f3b48c2335b674961db85be04d7a5eabccd27af