DAQ: Fix inserted TP length with active TTL channels
The function TP_GetTestPulseLengthInPoints returns the length of the TP
in points for the testpulse sampling interval.
In DC_PlaceDataInHardwareDataWave we adapt that length to represent the
TP length for DAQ taking the multiplier into account.

But taking only the multiplier into account is not enough as for ITC
hardware the presence of TTL channels influences the sampling interval
for DAQ but not for TP.

This results in the wrong length of the inserted TP which in this case
then also spills into the stimset data itself.

Fix that by introducing a mode flag to TP_GetTestPulseLengthInPoints
and use a generic solution which uses the factor from calling
DAP_GetSampInt for both modes.

Original-CommitID: ebb5ca6017b22a50f80f1ef11a10efd1d5d8e4b7