TP_Setup: Reset tick counters if required
Since da33596f (Rework TPStorage completely, 2018-10-18) we keep only
one TPStorage wave.

This change implied that we don't recreate TPStorage from scratch when
starting TP.

But for things like Autobias and dimension rescaling we use the seconds
since OS start to keep track of when we have to do them again. And this
of course breaks if you reuse an existing experiment from an earlier OS
start as then these timestamps are in the future.

Resetting them to zero if they are in the future fixes that.

Original-CommitID: 859d677e54a9e7b6d3f6c712eca96c6c3debc8ee