fix for decimationFactor < 1 for NI hardware resulting in OOR error
There was an issue in DC_PlaceDataInHardwareDataWave an index out of range
was accessed when a decimationFactor < 1 was in effect with a large data
wave.

The following example illustrates the problem:

Having a single active channel on NI allows to run at 500 kHz, whereas
the generated waves are at 200 kHz default, this results in a decimation factor
of 0.4.

For a stimSet of 720.000 points a data wave of 1.800.000 points is prepared.
When filling by auto indexing the data wave the last point reached by
0.4 * 1.799.999 is 719.999,6 which is beyond the last valid integer index
of the wave. (equals 720.000 - decimation factor)

It has been solved by limiting the source wave index to the last
valid integer index.

Original-CommitID: 29ae82a7fd4a19ff4e953b2e8ac5ed4d58969dfe