DC_MakeHardwareDataWave: Add dimension scaling to NI wave ref wave as well
For NI hardware the HardwareDataWave is a wave reference wave. This wave
is passed to the analysis functions via s.rawDACWave.

Add the sampling interval to the wave reference wave as well so that the
analysis functions still work.

Original-CommitID: e5dd459b6458cf1a83e05bb5763215606a936c9d