NWB_WriteStimsetTemplateWaves: Make more robust for partially missing parameter waves
The function WB_StimsetIsFromThirdParty returns one if one of the
parameter waves are missing, but we used to check that all are missing.

As we can not recreate a stimset if at least one parameter wave is
missing we can rewrite it in that way.

Original-CommitID: 7e3ebe4bd0e6619505ba8ac5a4fbbae60c2ce12b