WB_StimsetChildren: Handle non existing stimsets better
In c463da05 (fix for old stimset type, 2017-03-03) we added the check
that we return an empty list of child stimsts if the parameter waves
don't exist and the stimset itself exists.

This breaks when the stimset and its parameter waves are missing, as the
user for example deleted them.

Don't require that the stimset exists. The stimset existence check is
invalid anyway as we would need to recreate it instead of just checking
if the wave exists.

Original-CommitID: c94798b2848decfc33e0ac644c79ce253e948fee