NWB_WriteStimsetTemplateWaves: Handle deleted stimsets gracefully
If the user acquires data with a stimset and then deletes the stimset
and then tries to export into NWB we bug out.

As we can not do anything to get the deleted stimset back we now
complain only and continue.

Original-CommitID: 2182240b7be6944a817c89fc2c9e016a419914de