Included functionality of TP_StoreFullWave in SCOPE_UpdateOscilloscopeData
this removes also code duplication
Removed TP_StoreFullWave as it is no longer used

SCOPE_UpdateOscilloscopeData check for fifopos NaN if ITC finishes in DAQ mode

SCOPE_UpdateOscilloscopeData added ASSERT for the case tpStart < tpEnd

While this condition does not lead to a crash it indicates that
the new fifopos is smaller than the previous one.

Currently the tp loop skips handling in this case.
If this ASSERT comes up the origin should be reviewed.