Added call to SCOPE_UpdateGraph in Multi and Single Device Mode DAQ
A time interval constant DAQ_LIVE_UPDATE_INTERVAL similar to
TP_LIVE_UPDATE_INTERVAL is used to be able to tune the update interval.

UpdateGraph is only called when TP channels are present and
rescales axes of TP channels only. The data of the last TP defines the
new Y-scale.

The upper time axis is adjusted if at least one channel is in TP mode.