PA_ShowPulses: Make trace names maximally different
TraceNameList is really slow for large (>20k) number of traces as it
internally has complexity n^2.

Turns out, thanks to Adam Light from WaveMetrics for the hint, that if
you make the trace names maximally different from the start of the
string, TraceNameList is much faster.

Do that now here so to speed it up.

Original-CommitID: dc37e2fcdeb6bf382916d446271efaf217fbae66