SCOPE_UpdateOscilloscopeData: Use an integer as TPMarker
Since the introduction of threaded TP in 94358a04 (Added Structure for
input of asynchroneous in func TP_SendToAnalysis, 2018-12-06) we use a
measurement marker, drawn from a random distribution, to be used to
identify the testpulse.

For a future application we want to convert that marker to a string, to
avoid problems with doubles, we convert that double to a 32-bit integer.