PatchSeq/Rheobase: Make the DAScale stepsize configurable
In a future commit we want to change the stepsize depending on the
initial DAScale value.

So we need to store the initial DAScale stepsize in the labnotebook and
query that if we need it.

Add also some code in the tests to read it out.

Original-CommitID: 0315bd22dd3d1b0093e0686fe6145dbdbc7c7ae6