PatchSeq Rheobase: Search again with small DAScale values
Up to now we are finished once we have two sweeps which are
spike/non-spike. For small DAScale values this can be rather inaccurate
with a step size of 10pA.

Therefore we now use smaller steps, 2pA, and continue if we have met the
spike/non-spike condition with a low DAScale value.

Original-CommitID: a5d9c4475db304c85b06e7c50e58013584827024