BUG: Assert out during automated testing
The BUG() function was introduced to handle rare cases we know should
never be hit, and are thus a definite bug, but we still want to
continue.

But we definitly want to assert out during CI, so let's do it!

Original-CommitID: eb8754897a7db75b2722e66b0a8ddb6ae739bfbb