DAG_GetNumericalValue/DAG_GetTextualValue: Disambiguate the index parameter
We used to use the index parameter to signal that it is a unique control
or one with multiple entries.

This works in general but for the testing we want to pass in an index
which signals that this is a unique control.

Allow NaN to be used for that and introduce waveIndex for indexing the
GUiState wave.

Original-CommitID: 51f733a651c3a975e238976c00259c8af3d4e527