Introduce hardware type agnostic constants
Ever since the introduction of NI device support in fac268c1
(Acquisition support for NI DAC devices in multi device mode,
2018-08-03) we have relied also on the HARDWARE_ITC_MIN_SAMPINT_HZ
constant for the stimset frequency.

This is of course not correct anymore as it is only due to historic
reasons that the sampling interval of the stimset is the maximum sampling
interval for our ITC devices.

So let's introduce hardware agnostic constants.

Original-CommitID: a22dc301542415d1627763590118d1dde32ae9bc