TP during DAQ: Perform relevant TP Setup/Teardown
For TP during DAQ we want to also be able to use pressure methods.
We can not just call TP_Setup/TP_Teardown as that does other, TP only,
things as well.

Introduce TP_SetupCommon and TP_TeardownCommon and these can be called
if TP during DAQ channels are active as well.