AFH_GetAnalysisParam*: Tighten logic and add tests
We now assert out if the requested parameter name exists but does not
have the expected type. The behaviour that an non-existing parameter
returns successfully is not changed as we want to return a default value
in the future.

Original-CommitID: 3d9137b29f5b0f2de899fc0f32823dba49dd4934