PSQ_Rheobase: Handle DaScale of zero better
Currently we bug out if we encounter a DAScale of zero as in this
case we can not determine the pulse duration.

Preventing an initial DAScale value from a previous analysis function is
the easy part.

For the other cases where we use the stepsize to adapt the DAScale we
either fail due to limited resolution (2pA stepsize) or retry with
smaller values (10pA).

In addition we now also require that for spike/non-spike pattern we have
the same stepsize. For this feature we needed to introduce a new LBN
entry PSQ_FMT_LBN_STEPSIZE_FUTURE. This holds the step size we want to
use for future sweeps. In the POST_SWEEP_EVENT we write this into the
LBN using PSQ_FMT_LBN_STEPSIZE as that was the step size which was
really used for that sweep. This solution is required that we can
distinguish between "should be used from now on" and "was used for this
sweep".

Original-CommitID: 5bdb761f5843df91a9d506a1880a1a55c7b7d574