GetChannelClampMode: Extend to also hold headstage information
It is currently suprisingly difficult to get the headstage from a DA/AD
channel number.

The old approach involved ChanAmpAssign and the ChannelClampMode waves
and was slow. This is ugly especially when called during the mid sweep
event.

As ChannelClampMode is one of the few waves which is indexed by
channel number, the obvious solution is to extend it to hold the
headstage as well.

For wave upgrades we prefill the headstage numbers using the old
algorithm.

Original-CommitID: 134392ccd47ccee801593250561a60e74a842e0c