DC_PlaceDataInHardwareDataWave: Use zero as TP amplitude for unassociated DACs
Since forever we are just using the clamp mode from ChannelClampMode for
the DA channel.

But that was prior to 46da8c4e (GetChannelClampMode: Initialize wave to
NaN, 2019-03-08) initialized to zero. So for unassociated DACs, which
never have a clamp mode, we used TP ampltitude as for V_CLAMP_MODE (aka 0).

With NaN initialization this now bugs out.

Use zero as TP amplitude for unassociated DACs.

Original-CommitID: 074281cdd988b0beed7496145ea71651627a75be