PSQ_CalculateAvg: Fixup range of points for NI hardware and sampling interval multiplier > 1
The analysis function PSQ_DAScale checks that the stimset is long enough
in the PRE_DAQ/PRE_SET events.

But the stimset length is with the full resolution and aquisition with
NI hardware and a sampling interval multiplier can result in the last
point of the HardwareDataWave not being present.

See b96781ca (fix for decimationFactor < 1 for NI hardware resulting in OOR
error, 2018-11-30) and 0de87ea6 (fix: changed calculation of real sample rate wave sizes,
2018-08-08) for the reasons why we choose to make the HardwareDataWave
one element smaller.

The fix in PSQ_CalculateAvg is also safe as we only call it for a valid
time range.

Original-CommitID: 04f3f877559975780c07789b739529144d613058