NI Hardware: Make FIFO size dependent on Igor Pro bitness
In c80e7741 (NI hardware: Raise fifo size to 120s, 2019-01-29) we
increased the fifo size for being able to debug issues, as we don't want
the fifo to overflow.

But the new size is too large for Igor Pro 32bit and results in the fact
that FifoStatus always returns V_FIFOChunks == 0.

Let's special case the size to the old value in the 32bit case.

Original-CommitID: a82b562df58c1d388a873891f080e90bd3969b4d